renamed 3scale_manaement_api to 3scale_management_api
